reorg of the class into the new package hierarchy after rebase
