+++ src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/BgpAst.java
@@ -1,4 +1,4 @@
-package fr.inria.corese.core.next.query.sparql.ast;
+package fr.inria.corese.core.next.query.api.sparql.ast;
 
 import java.util.List;
