+++ src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/PatternAst.java
@@ -1,4 +1,4 @@
-package fr.inria.corese.core.next.query.sparql.ast;
+package fr.inria.corese.core.next.query.api.sparql.ast;
 
 /**
  * Element of a group graph pattern (BGP, optional, union, etc.).
