+++ src/main/java/fr/inria/corese/core/next/query/api/sparql/ast/TermAst.java
@@ -1,4 +1,4 @@
-package fr.inria.corese.core.next.query.sparql.ast;
+package fr.inria.corese.core.next.query.api.sparql.ast;
 
 /**
  * AST node for a SPARQL term (variable, IRI or literal) in a triple pattern.
