+++ src/main/java/fr/inria/corese/core/next/query/api/sparql/options/BaseIRIOptions.java
@@ -1,4 +1,4 @@
-package fr.inria.corese.core.next.query.sparql.options;
+package fr.inria.corese.core.next.query.api.sparql.options;
 
 public interface BaseIRIOptions {
     String getBaseIRI();
