--- src/main/java/fr/inria/corese/core/next/query/impl/sparql/ast/SelectQueryAst.java
@@ -1,4 +1,0 @@
-package fr.inria.corese.core.next.query.impl.sparql.ast;
-
-public class SelectQueryAst {
-}
