+++ src/test/java/fr/inria/corese/core/next/query/impl/parser/SparqlAstBuilderTest.java
@@ -1,7 +1,7 @@
-package fr.inria.corese.core.next.query.sparql.parser;
+package fr.inria.corese.core.next.query.impl.parser;
 
 
-import fr.inria.corese.core.next.query.sparql.ast.*;
+import fr.inria.corese.core.next.query.api.sparql.ast.*;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
