+++ src/test/java/fr/inria/corese/core/next/query/impl/parser/SparqlParserBgpTest.java
@@ -1,8 +1,8 @@
-package fr.inria.corese.core.next.query.sparql.parser;
+package fr.inria.corese.core.next.query.impl.parser;
 
-import fr.inria.corese.core.next.data.api.query.exception.QueryException;
-import fr.inria.corese.core.next.data.api.query.exception.QuerySyntaxException;
-import fr.inria.corese.core.next.query.sparql.ast.*;
+import fr.inria.corese.core.next.query.api.exception.QueryException;
+import fr.inria.corese.core.next.query.api.exception.QuerySyntaxException;
+import fr.inria.corese.core.next.query.api.sparql.ast.*;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
