+++ src/test/java/fr/inria/corese/core/next/query/impl/parser/SparqlParserTest.java
@@ -1,10 +1,10 @@
-package fr.inria.corese.core.next.query.sparql.parser;
+package fr.inria.corese.core.next.query.impl.parser;
 
-import fr.inria.corese.core.next.data.api.query.exception.QueryException;
-import fr.inria.corese.core.next.query.QueryOptions;
-import fr.inria.corese.core.next.query.sparql.ast.BgpAst;
-import fr.inria.corese.core.next.query.sparql.ast.QueryAst;
-import fr.inria.corese.core.next.query.sparql.ast.VarAst;
+import fr.inria.corese.core.next.query.api.exception.QueryException;
+import fr.inria.corese.core.next.query.api.io.parser.QueryOptions;
+import fr.inria.corese.core.next.query.api.sparql.ast.BgpAst;
+import fr.inria.corese.core.next.query.api.sparql.ast.QueryAst;
+import fr.inria.corese.core.next.query.api.sparql.ast.VarAst;
 import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
